enhancement: hot tier enabled for standalone mode
